move batch string interface
