+++ csharp/sources/Valkey.Glide/Pipeline/BaseBatch.StringCommands.cs
@@ -1,6 +1,5 @@
 ﻿// Copyright Valkey GLIDE Project Contributors - SPDX Identifier: Apache-2.0
 
-using Valkey.Glide.Commands;
 using Valkey.Glide.Internals;
 
 namespace Valkey.Glide.Pipeline;
